rbd: use rbd snapshots and cloning for ephemeral disk snapshots
When rbd is used for ephemeral disks and image storage, we can avoid
local disk by doing the snapshot with ceph, and sending the location
of it to glance.

At this point glance is effectively the owner of the snapshot, and
nova cannot delete the original disk, so a corresponding change to
glance is needed to handle this cleanup later, when the snapshot is
deleted.

To accomplish this, create an rbd clone and store the clone's location
in glance, and when nova deletes an rbd disk with snapshots, rename
the disk to include a marker that glance looks for to know they can be
deleted when glance-scrubber gets to them.

This requires changing the typical permissions for the glance ceph
user to allow writing to the pool where vm images are stored.

Signed-off-by: Josh Durgin <josh.durgin@inktank.com>
Closes-rally-bug: DE1186
Implements: rally-task TA8264
Not-in-upstream: true